rename core_vendor_information.js to core_vendor_lists.js
