+++ src/scripts/core/core_command_collection.js
@@ -1,7 +1,7 @@
 import { logError, logInfo } from './core_log';
 import { getCommandCollection } from './core_utils';
 import { getConsentDataString, getVendorConsentData, getPublisherConsentData } from './core_consents';
-import { getVendorList, loadVendorListAndCustomVendorList } from './core_vendor_information';
+import { getVendorList, loadVendorListAndCustomVendorList } from './core_vendor_lists';
 
 const commands = {
   getVendorConsents: (vendorIds) => {
