+++ src/scripts/core/core_consents.js
@@ -1,6 +1,6 @@
 import {getSoiCookie} from './core_cookies';
 import {getCustomPurposeIds, gdprApplies} from './core_config';
-import {getLimitedVendorIds, getPurposeIds} from './core_vendor_information';
+import {getLimitedVendorIds, getPurposeIds} from './core_vendor_lists';
 import {OIL_SPEC} from './core_constants';
 
 export function getVendorConsentData(vendorIds) {
