+++ src/scripts/core/core_tag_management.js
@@ -1,7 +1,7 @@
 import { MANAGED_TAG_IDENTIFIER, MANAGED_TAG_IDENTIFIER_ATTRIBUTE, MANAGED_TAG_PURPOSES_ATTRIBUTE } from './core_constants';
 import { getSoiCookie } from './core_cookies';
 import { arrayContainsArray } from './core_utils';
-import { getPurposeIds } from './core_vendor_information';
+import { getPurposeIds } from './core_vendor_lists';
 import { getCustomPurposeIds } from './core_config';
 
 export function manageDomElementActivation() {
