+++ src/scripts/core/core_vendor_lists.js
