+++ src/scripts/userview/view/oil.advanced.settings.standard.js
@@ -1,18 +1,12 @@
 import '../../../styles/cpc_standard.scss';
-import { OIL_LABELS } from '../userview_constants.js';
-import { forEach } from '../userview_modal.js';
-import { getLabel, getLabelWithDefault, getTheme } from '../userview_config.js';
-import { getCustomPurposes } from '../../core/core_config.js';
-import { JS_CLASS_BUTTON_OPTIN, OIL_GLOBAL_OBJECT_NAME } from '../../core/core_constants.js';
-import { setGlobalOilObject } from '../../core/core_utils.js';
-import {
-  getPurposes,
-  getVendorList,
-  getCustomVendorList,
-  getVendorsToDisplay
-} from '../../core/core_vendor_information.js';
-import { BackButton, YesButton } from './components/oil.buttons.js';
-import { getCustomVendorListUrl } from '../../core/core_config';
+import { OIL_LABELS } from '../userview_constants';
+import { forEach } from '../userview_modal';
+import { getLabel, getLabelWithDefault, getTheme } from '../userview_config';
+import { getCustomPurposes, getCustomVendorListUrl } from '../../core/core_config';
+import { JS_CLASS_BUTTON_OPTIN, OIL_GLOBAL_OBJECT_NAME } from '../../core/core_constants';
+import { setGlobalOilObject } from '../../core/core_utils';
+import { getCustomVendorList, getPurposes, getVendorList, getVendorsToDisplay } from '../../core/core_vendor_lists';
+import { BackButton, YesButton } from './components/oil.buttons';
 
 
 const CLASS_NAME_FOR_ACTIVE_MENU_SECTION = 'as-oil-cpc__category-link--active';
