+++ src/scripts/userview/view/oil.advanced.settings.tabs.js
@@ -1,13 +1,11 @@
 import '../../../styles/cpc_tabs.scss';
-import { OIL_LABELS } from '../userview_constants.js';
-import { forEach } from '../userview_modal.js';
-import { getLabel, getLabelWithDefault } from '../userview_config.js';
-import { getCustomPurposes } from '../../core/core_config.js';
-import { JS_CLASS_BUTTON_OPTIN, OIL_GLOBAL_OBJECT_NAME } from '../../core/core_constants.js';
-import { getPurposes, getVendorList, getVendorsToDisplay } from '../../core/core_vendor_information.js';
-import { BackButton, YesButton } from './components/oil.buttons.js';
-import { getCustomVendorList } from '../../core/core_vendor_information';
-import { getCustomVendorListUrl } from '../../core/core_config';
+import { OIL_LABELS } from '../userview_constants';
+import { forEach } from '../userview_modal';
+import { getLabel, getLabelWithDefault } from '../userview_config';
+import { getCustomPurposes, getCustomVendorListUrl } from '../../core/core_config';
+import { JS_CLASS_BUTTON_OPTIN, OIL_GLOBAL_OBJECT_NAME } from '../../core/core_constants';
+import { getCustomVendorList, getPurposes, getVendorList, getVendorsToDisplay } from '../../core/core_vendor_lists';
+import { BackButton, YesButton } from './components/oil.buttons';
 
 export function oilAdvancedSettingsTemplate() {
   return `
