+++ test/specs/core/core_consents.spec.js
@@ -1,7 +1,7 @@
 import { getConsentDataString, getPublisherConsentData, getVendorConsentData } from '../../../src/scripts/core/core_consents';
 import * as CoreCookies from '../../../src/scripts/core/core_cookies';
 import * as CoreConfig from '../../../src/scripts/core/core_config';
-import * as CoreVendorInformation from '../../../src/scripts/core/core_vendor_information';
+import * as CoreVendorInformation from '../../../src/scripts/core/core_vendor_lists';
 import { OIL_SPEC } from '../../../src/scripts/core/core_constants';
 import { resetOil } from '../../test-utils/utils_reset';
