+++ test/specs/core/core_vendor_information.spec.js
@@ -14,12 +14,12 @@ import {
   cachedVendorList,
   pendingVendorListPromise,
   getCustomVendorList
-} from '../../../src/scripts/core/core_vendor_information';
+} from '../../../src/scripts/core/core_vendor_lists';
 import VENDOR_LIST from '../../fixtures/vendorlist/simple_vendor_list.json';
 import CUSTOM_VENDOR_LIST from '../../fixtures/vendorlist/custom_vendor_list.json';
 import { resetOil } from '../../test-utils/utils_reset';
 
-describe('core_vendor_information', () => {
+describe('core_vendor_lists', () => {
 
   const WHITELISTED_VENDORS = [1, 2];
   const BLACKLISTED_VENDORS = Array.apply(null, {length: (DEFAULT_VENDOR_LIST.maxVendorId - 2)}).map(Number.call, Number).slice(1);
